+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * **Breaking**(In case you have implemented a custom cart service): Extend the cart service interface with `UpdateItemBundleConfig` to allow updating bundles that have already been placed inside the cart.
 * GraphQL:
   * Add new mutation `Commerce_Cart_UpdateItemBundleConfig` to update bundle configs for existing cart items
+  * Expose selected bundle configuration on the cart items 
 
 **checkout**
 * initialize place order metrics with 0 on application start to follow prometheus best practices